+++ cmsat/sccfinder.cpp
@@ -188,9 +188,9 @@ void SCCFinder::tarjan(const uint32_t vertex)
     }
 }
 
-uint64_t SCCFinder::memUsed() const
+size_t SCCFinder::memUsed() const
 {
-    uint64_t mem = 0;
+    size_t mem = 0;
     mem += index.capacity()*sizeof(uint32_t);
     mem += lowlink.capacity()*sizeof(uint32_t);
     mem += stack.size()*sizeof(uint32_t); //TODO under-estimates
