+++ cmsat/simplifier.h
@@ -100,8 +100,8 @@ public:
         , const vector<uint32_t>& interToOuter
     );
     bool unEliminate(const Var var);
-    uint64_t memUsed() const;
-    uint64_t memUsedXor() const;
+    size_t memUsed() const;
+    size_t memUsedXor() const;
 
     //UnElimination
     void print_blocked_clauses_reverse() const;
